chore: 다이얼로그를 한 패키지에 모음
