+++ android/app/src/main/java/com/now/naaga/presentation/common/dialog/CameraPermissionDialog.kt
@@ -1,8 +1,6 @@
-package com.now.naaga.presentation.beginadventure
+package com.now.naaga.presentation.common.dialog
 
-import android.Manifest.permission.ACCESS_COARSE_LOCATION
 import android.content.Intent
-import android.content.pm.PackageManager.PERMISSION_GRANTED
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.net.Uri
@@ -11,25 +9,23 @@ import android.provider.Settings
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat.checkSelfPermission
 import androidx.fragment.app.DialogFragment
-import com.now.naaga.R
 import com.now.naaga.data.firebase.analytics.AnalyticsDelegate
+import com.now.naaga.data.firebase.analytics.CAMERA_PERMISSION_OPEN_SETTING
 import com.now.naaga.data.firebase.analytics.DefaultAnalyticsDelegate
-import com.now.naaga.data.firebase.analytics.LOCATION_PERMISSION_OPEN_SETTING
-import com.now.naaga.databinding.DialogLocationPermissionBinding
+import com.now.naaga.databinding.DialogCameraPermissionBinding
 import com.now.naaga.util.dpToPx
 import com.now.naaga.util.getWidthProportionalToDevice
 
-class LocationPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAnalyticsDelegate() {
-    private lateinit var binding: DialogLocationPermissionBinding
+class CameraPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAnalyticsDelegate() {
+    private lateinit var binding: DialogCameraPermissionBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        binding = DialogLocationPermissionBinding.inflate(layoutInflater)
+        binding = DialogCameraPermissionBinding.inflate(layoutInflater)
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
         return binding.root
     }
@@ -41,18 +37,17 @@ class LocationPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultA
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        checkPermission()
         setSize()
         binding.btnDialogLocationSetting.setOnClickListener {
-            logClickEvent(requireContext().getViewEntryName(it), LOCATION_PERMISSION_OPEN_SETTING)
+            logClickEvent(requireContext().getViewEntryName(it), CAMERA_PERMISSION_OPEN_SETTING)
             openSetting()
             dismiss()
         }
     }
 
     private fun setSize() {
-        val dialogWidth = getWidthProportionalToDevice(requireContext(), WIDTH_RATE)
-        val dialogHeight = dpToPx(requireContext(), HEIGHT)
+        val dialogWidth = getWidthProportionalToDevice(requireContext(), LocationPermissionDialog.WIDTH_RATE)
+        val dialogHeight = dpToPx(requireContext(), LocationPermissionDialog.HEIGHT)
         dialog?.window?.setLayout(dialogWidth, dialogHeight)
     }
 
@@ -64,27 +59,7 @@ class LocationPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultA
         startActivity(appDetailsIntent)
     }
 
-    private fun checkPermission() {
-        if (checkSelfPermission(requireContext(), ACCESS_COARSE_LOCATION) == PERMISSION_GRANTED) {
-            setDescription(true)
-            return
-        }
-        setDescription(false)
-    }
-
-    private fun setDescription(isApproximateAccessGranted: Boolean) {
-        val description: String = if (isApproximateAccessGranted) {
-            getString(R.string.locationDialog_approximate_description)
-        } else {
-            getString(R.string.locationDialog_description)
-        }
-
-        binding.tvDialogLocationDescription.text = description
-    }
-
     companion object {
-        const val WIDTH_RATE = 0.83f
-        const val HEIGHT = 400
-        const val TAG_LOCATION_DIALOG = "LOCATION"
+        const val TAG_CAMERA_DIALOG = "CAMERA"
     }
 }
