+++ android/app/src/main/java/com/now/naaga/presentation/common/dialog/LocationPermissionDialog.kt
@@ -1,6 +1,8 @@
-package com.now.naaga.presentation.upload
+package com.now.naaga.presentation.common.dialog
 
+import android.Manifest.permission.ACCESS_COARSE_LOCATION
 import android.content.Intent
+import android.content.pm.PackageManager.PERMISSION_GRANTED
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.net.Uri
@@ -9,24 +11,25 @@ import android.provider.Settings
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.content.ContextCompat.checkSelfPermission
 import androidx.fragment.app.DialogFragment
+import com.now.naaga.R
 import com.now.naaga.data.firebase.analytics.AnalyticsDelegate
-import com.now.naaga.data.firebase.analytics.CAMERA_PERMISSION_OPEN_SETTING
 import com.now.naaga.data.firebase.analytics.DefaultAnalyticsDelegate
-import com.now.naaga.databinding.DialogCameraPermissionBinding
-import com.now.naaga.presentation.beginadventure.LocationPermissionDialog
+import com.now.naaga.data.firebase.analytics.LOCATION_PERMISSION_OPEN_SETTING
+import com.now.naaga.databinding.DialogLocationPermissionBinding
 import com.now.naaga.util.dpToPx
 import com.now.naaga.util.getWidthProportionalToDevice
 
-class CameraPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAnalyticsDelegate() {
-    private lateinit var binding: DialogCameraPermissionBinding
+class LocationPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAnalyticsDelegate() {
+    private lateinit var binding: DialogLocationPermissionBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
-        binding = DialogCameraPermissionBinding.inflate(layoutInflater)
+        binding = DialogLocationPermissionBinding.inflate(layoutInflater)
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
         return binding.root
     }
@@ -38,17 +41,18 @@ class CameraPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAna
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        checkPermission()
         setSize()
         binding.btnDialogLocationSetting.setOnClickListener {
-            logClickEvent(requireContext().getViewEntryName(it), CAMERA_PERMISSION_OPEN_SETTING)
+            logClickEvent(requireContext().getViewEntryName(it), LOCATION_PERMISSION_OPEN_SETTING)
             openSetting()
             dismiss()
         }
     }
 
     private fun setSize() {
-        val dialogWidth = getWidthProportionalToDevice(requireContext(), LocationPermissionDialog.WIDTH_RATE)
-        val dialogHeight = dpToPx(requireContext(), LocationPermissionDialog.HEIGHT)
+        val dialogWidth = getWidthProportionalToDevice(requireContext(), WIDTH_RATE)
+        val dialogHeight = dpToPx(requireContext(), HEIGHT)
         dialog?.window?.setLayout(dialogWidth, dialogHeight)
     }
 
@@ -60,7 +64,27 @@ class CameraPermissionDialog : DialogFragment(), AnalyticsDelegate by DefaultAna
         startActivity(appDetailsIntent)
     }
 
+    private fun checkPermission() {
+        if (checkSelfPermission(requireContext(), ACCESS_COARSE_LOCATION) == PERMISSION_GRANTED) {
+            setDescription(true)
+            return
+        }
+        setDescription(false)
+    }
+
+    private fun setDescription(isApproximateAccessGranted: Boolean) {
+        val description: String = if (isApproximateAccessGranted) {
+            getString(R.string.locationDialog_approximate_description)
+        } else {
+            getString(R.string.locationDialog_description)
+        }
+
+        binding.tvDialogLocationDescription.text = description
+    }
+
     companion object {
-        const val TAG_CAMERA_DIALOG = "CAMERA"
+        const val WIDTH_RATE = 0.83f
+        const val HEIGHT = 400
+        const val TAG_LOCATION_DIALOG = "LOCATION"
     }
 }
