+++ android/app/src/main/java/com/now/naaga/presentation/setting/SettingActivity.kt
@@ -9,8 +9,8 @@ import androidx.appcompat.app.AppCompatActivity
 import com.now.naaga.R
 import com.now.naaga.data.throwable.DataThrowable
 import com.now.naaga.databinding.ActivitySettingBinding
+import com.now.naaga.presentation.common.dialog.NaagaAlertDialog
 import com.now.naaga.presentation.login.LoginActivity
-import com.now.naaga.presentation.onadventure.NaagaAlertDialog
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
