+++ android/app/src/main/java/com/now/naaga/presentation/upload/UploadActivity.kt
@@ -27,9 +27,10 @@ import com.now.naaga.data.firebase.analytics.DefaultAnalyticsDelegate
 import com.now.naaga.data.firebase.analytics.UPLOAD_OPEN_CAMERA
 import com.now.naaga.data.throwable.DataThrowable
 import com.now.naaga.databinding.ActivityUploadBinding
-import com.now.naaga.presentation.beginadventure.LocationPermissionDialog
-import com.now.naaga.presentation.beginadventure.LocationPermissionDialog.Companion.TAG_LOCATION_DIALOG
-import com.now.naaga.presentation.upload.CameraPermissionDialog.Companion.TAG_CAMERA_DIALOG
+import com.now.naaga.presentation.common.dialog.CameraPermissionDialog
+import com.now.naaga.presentation.common.dialog.CameraPermissionDialog.Companion.TAG_CAMERA_DIALOG
+import com.now.naaga.presentation.common.dialog.LocationPermissionDialog
+import com.now.naaga.presentation.common.dialog.LocationPermissionDialog.Companion.TAG_LOCATION_DIALOG
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -93,7 +94,11 @@ class UploadActivity : AppCompatActivity(), AnalyticsDelegate by DefaultAnalytic
                     changeVisibility(binding.lottieUploadLoading, View.GONE)
                     finish()
                 }
-                UploadStatus.PENDING -> { changeVisibility(binding.lottieUploadLoading, View.VISIBLE) }
+
+                UploadStatus.PENDING -> {
+                    changeVisibility(binding.lottieUploadLoading, View.VISIBLE)
+                }
+
                 UploadStatus.FAIL -> {
                     changeVisibility(binding.lottieUploadLoading, View.GONE)
                     shortToast(getString(R.string.upload_fail_submit))
@@ -105,18 +110,27 @@ class UploadActivity : AppCompatActivity(), AnalyticsDelegate by DefaultAnalytic
                 UploadViewModel.ERROR_STORE_PHOTO -> {
                     shortToast(getString(R.string.upload_error_store_photo_message))
                 }
+
                 UploadViewModel.ALREADY_EXISTS_NEARBY -> {
                     shortToast(getString(R.string.upload_error_already_exists_nearby_message))
                 }
-                UploadViewModel.ERROR_POST_BODY -> { shortToast(getString(R.string.upload_error_post_message)) }
+
+                UploadViewModel.ERROR_POST_BODY -> {
+                    shortToast(getString(R.string.upload_error_post_message))
+                }
             }
         }
     }
 
     private fun changeVisibility(view: View, status: Int) {
         when (status) {
-            View.VISIBLE -> { view.visibility = status }
-            View.GONE -> { view.visibility = status }
+            View.VISIBLE -> {
+                view.visibility = status
+            }
+
+            View.GONE -> {
+                view.visibility = status
+            }
         }
     }
